+++ pkgs/native_toolchain_c/lib/src/cbuilder/cbuilder.dart
@@ -20,10 +20,17 @@ class Language {
   const Language._(this.name);
 
   static const Language c = Language._('c');
+
   static const Language cpp = Language._('c++');
 
+  static const Language objectiveC = Language._('objective c');
+
   /// Known values for [Language].
-  static const List<Language> values = [c, cpp];
+  static const List<Language> values = [
+    c,
+    cpp,
+    objectiveC,
+  ];
 
   @override
   String toString() => name;
@@ -62,6 +69,20 @@ class CBuilder implements Builder {
   ///
   /// Used to output the [BuildOutput.dependencies].
   final List<String> includes;
+
+  /// Frameworks to link.
+  ///
+  /// Only effective if [language] is [Language.objectiveC].
+  ///
+  /// Defaults to `['Foundation']`.
+  ///
+  /// Not used to output the [BuildOutput.dependencies], frameworks can be
+  /// mentioned by name if they are available on the system, so the file path
+  /// is not known. If you're depending on your own frameworks add them to
+  /// [BuildOutput.dependencies] manually.
+  final List<String> frameworks;
+
+  static const List<String> _defaultFrameworks = ['Foundation'];
 
   /// The dart files involved in building this artifact.
   ///
@@ -155,6 +176,7 @@ class CBuilder implements Builder {
     required this.assetName,
     this.sources = const [],
     this.includes = const [],
+    this.frameworks = _defaultFrameworks,
     required this.dartBuildFiles,
     @visibleForTesting this.installName,
     this.flags = const [],
@@ -172,6 +194,7 @@ class CBuilder implements Builder {
     required this.name,
     this.sources = const [],
     this.includes = const [],
+    this.frameworks = _defaultFrameworks,
     required this.dartBuildFiles,
     this.flags = const [],
     this.defines = const {},
@@ -221,6 +244,7 @@ class CBuilder implements Builder {
         logger: logger,
         sources: sources,
         includes: includes,
+        frameworks: frameworks,
         dynamicLibrary: _type == _CBuilderType.library &&
                 linkMode == DynamicLoadingBundled()
             ? libUri