+++ pkgs/native_toolchain_c/lib/src/cbuilder/run_cbuilder.dart
@@ -23,6 +23,7 @@ class RunCBuilder {
   final Logger? logger;
   final List<Uri> sources;
   final List<Uri> includes;
+  final List<String> frameworks;
   final Uri? executable;
   final Uri? dynamicLibrary;
   final Uri? staticLibrary;
@@ -47,6 +48,7 @@ class RunCBuilder {
     this.logger,
     this.sources = const [],
     this.includes = const [],
+    required this.frameworks,
     this.executable,
     this.dynamicLibrary,
     this.staticLibrary,
@@ -246,6 +248,12 @@ class RunCBuilder {
           if (value == null) '-D$name' else '-D$name=$value',
         for (final include in includes) '-I${include.toFilePath()}',
         ...sourceFiles,
+        if (language == Language.objectiveC) ...[
+          for (final framework in frameworks) ...[
+            '-framework',
+            framework,
+          ],
+        ],
         if (executable != null) ...[
           '-o',
           outDir.resolveUri(executable!).toFilePath(),